Establish test helper functions to avoid code duplication
