+++ setup.cfg
@@ -1,6 +1,12 @@
 [flake8]
 max-line-length = 130
 extend-ignore = E203
+per-file-ignores =
+    # imported but unused
+    test_s3.py: F401, F811
 
 [tool:pytest]
 addopts = --cov=ils_middleware --cov-report=xml --cov-report=term
+
+[pytest]
+norecursedirs=tests/helpers
