+++ tests/conftest.py
@@ -1,0 +1,4 @@
+import sys
+import os
+
+sys.path.append(os.path.join(os.path.dirname(__file__), "helpers"))
